trello-service-2/requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-cloud-storage
google-cloud-pubsub